Token required for deletion

+++ controllers/blog.js
@@ -36,6 +36,11 @@ blogsRouter.get("/:id", async (request, response) => {
 });
 
 blogsRouter.delete("/:id", async (request, response) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+
+  if (!decodedToken.id)
+    return response.status(401).json({ error: "invalid token" });
+
   await Blog.findByIdAndRemove(request.params.id);
 
   response.status(204).end();
